BillboardList(VM + Model) GameObjectEditor(Unfinished View)

+++ MapEditor/Editor/GameEntities/Vector3.cs
@@ -2,11 +2,11 @@
 {
     public struct Vector3
     {
-        public float X;
-        public float Y;
-        public float Z;
+        public double X;
+        public double Y;
+        public double Z;
 
-        public Vector3(float x, float y, float z)
+        public Vector3(double x, double y, double z)
         {
             X = x;
             Y = y;
